move slicing code from core.py to slicing.py
