+++ dask/array/core.py
@@ -347,429 +347,6 @@ def rec_concatenate(arrays, axis=0):
     return np.concatenate(arrays, axis=axis)
 
 
-def new_blockdim(dim_shape, lengths, index):
-    """
-
-    >>> new_blockdim(100, [20, 10, 20, 10, 40], slice(0, 90, 2))
-    [10, 5, 10, 5, 15]
-
-    >>> new_blockdim(100, [20, 10, 20, 10, 40], [5, 1, 30, 22])
-    [4]
-    """
-    if isinstance(index, list):
-        return [len(index)]
-    assert not isinstance(index, int)
-    pairs = sorted(_slice_1d(dim_shape, lengths, index).items(), key=first)
-    return [(slc.stop - slc.start) // slc.step for _, slc in pairs]
-
-
-def _slice_1d(dim_shape, lengths, index):
-    """Returns a dict of {blocknum: slice}
-
-    This function figures out where each slice should start in each
-    block for a single dimension. If the slice won't return any elements
-    in the block, that block will not be in the output.
-
-    Parameters
-    ----------
-
-    dim_shape - the number of elements in this dimension.
-      This should be a positive, non-zero integer
-    blocksize - the number of elements per block in this dimension
-      This should be a positive, non-zero integer
-    index - a description of the elements in this dimension that we want
-      This might be an integer, a slice(), or an Ellipsis
-
-    Returns
-    -------
-
-    dictionary where the keys are the integer index of the blocks that
-      should be sliced and the values are the slices
-
-    Examples
-    --------
-
-    100 length array cut into length 20 pieces, slice 0:35
-
-    >>> _slice_1d(100, [20, 20, 20, 20, 20], slice(0, 35))
-    {0: slice(0, 20, 1), 1: slice(0, 15, 1)}
-
-    Support irregular blocks and various slices
-
-    >>> _slice_1d(100, [20, 10, 10, 10, 25, 25], slice(10, 35))
-    {0: slice(10, 20, 1), 1: slice(0, 10, 1), 2: slice(0, 5, 1)}
-
-    Support step sizes
-
-    >>> _slice_1d(100, [15, 14, 13], slice(10, 41, 3))
-    {0: slice(10, 15, 3), 1: slice(1, 14, 3), 2: slice(2, 12, 3)}
-
-    >>> _slice_1d(100, [20, 20, 20, 20, 20], slice(0, 100, 40))  # step > blocksize
-    {0: slice(0, 20, 40), 2: slice(0, 20, 40), 4: slice(0, 20, 40)}
-
-    Also support indexing single elements
-
-    >>> _slice_1d(100, [20, 20, 20, 20, 20], 25)
-    {1: 5}
-
-    And negative slicing
-
-    >>> _slice_1d(100, [20, 20, 20, 20, 20], slice(100, 0, -3))
-    {0: slice(-2, -20, -3), 1: slice(-1, -21, -3), 2: slice(-3, -21, -3), 3: slice(-2, -21, -3), 4: slice(-1, -21, -3)}
-
-    >>> _slice_1d(100, [20, 20, 20, 20, 20], slice(100, 12, -3))
-    {0: slice(-2, -8, -3), 1: slice(-1, -21, -3), 2: slice(-3, -21, -3), 3: slice(-2, -21, -3), 4: slice(-1, -21, -3)}
-
-    >>> _slice_1d(100, [20, 20, 20, 20, 20], slice(100, -12, -3))
-    {4: slice(-1, -12, -3)}
-    """
-    if isinstance(index, int):
-        i = 0
-        ind = index
-        lens = list(lengths)
-        while ind >= lens[0]:
-            i += 1
-            ind -= lens.pop(0)
-        return {i: ind}
-
-    assert isinstance(index, slice)
-
-    step = index.step or 1
-    if step > 0:
-        start = index.start or 0
-        stop = index.stop or dim_shape
-    else:
-        start = index.start or dim_shape - 1
-        start = dim_shape - 1 if start >= dim_shape else start
-        stop = -(dim_shape + 1) if index.stop is None else index.stop
-
-    if start < 0:
-        start += dim_shape
-    if stop < 0:
-        stop += dim_shape
-
-    d = dict()
-    if step > 0:
-        for i, length in enumerate(lengths):
-            if start < length and stop > 0:
-                d[i] = slice(start, min(stop, length), step)
-                start = (start - length) % step
-            else:
-                start = start - length
-            stop -= length
-    else:
-        stop -= dim_shape
-        tail_index = list(accumulate(add, lengths))
-        pos_step = abs(step) # 11%3==2, 11%-3==-1. Need positive step for %
-
-        offset = 0
-        for i, length in zip(range(len(lengths)-1, -1, -1), reversed(lengths)):
-            if start + length >= tail_index[i] and stop < 0:
-                d[i] = slice(start - tail_index[i],
-                             max(stop, -length - 1), step)
-                # The offset accumulates over time from the start point
-                offset = (offset + pos_step - (length % pos_step)) % pos_step
-                start = tail_index[i] - 1 - length - offset
-
-            stop += length
-
-    return d
-
-
-def partition_by_size(sizes, seq):
-    """
-
-    >>> partition_by_size([10, 20, 10], [1, 5, 9, 12, 29, 35])
-    [[1, 5, 9], [2, 19], [5]]
-    """
-    seq = list(seq)
-    pretotal = 0
-    total = 0
-    i = 0
-    result = list()
-    for s in sizes:
-        total += s
-        L = list()
-        while i < len(seq) and seq[i] < total:
-            L.append(seq[i] - pretotal)
-            i += 1
-        result.append(L)
-        pretotal += s
-    return result
-
-
-def take(outname, inname, blockdims, index, axis=0):
-    """ Index array with an iterable of index
-
-    Mimics ``np.take``
-
-    >>> take('y', 'x', [(20, 20, 20, 20)], [5, 1, 47, 3], axis=0)  # doctest: +SKIP
-    {('y', 0): (getitem, (np.concatenate, [(getitem, ('x', 0), ([1, 3, 5],)),
-                                           (getitem, ('x', 2), ([7],))],
-                                          0),
-                         (2, 0, 4, 1))}
-    """
-    n = len(blockdims)
-    sizes = blockdims[axis]  # the blocksizes on the axis that we care about
-
-    index_lists = partition_by_size(sizes, sorted(index))
-
-    dims = [[0] if axis == i else list(range(len(bd)))
-                for i, bd in enumerate(blockdims)]
-    keys = list(product([outname], *dims))
-
-    colon = slice(None, None, None)
-
-    rev_index = tuple(map(sorted(index).index, index))
-    vals = [(getitem, (np.concatenate,
-                  (list, [(getitem, ((inname,) + d[:axis] + (i,) + d[axis+1:]),
-                                   ((colon,)*axis + (IL,) + (colon,)*(n-axis-1)))
-                                for i, IL in enumerate(index_lists)
-                                if IL]),
-                        axis),
-                     ((colon,)*axis + (rev_index,) + (colon,)*(n-axis-1)))
-            for d in product(*dims)]
-
-    return dict(zip(keys, vals))
-
-
-def posify_index(shape, ind):
-    """
-
-    >>> posify_index(10, 3)
-    3
-    >>> posify_index(10, -3)
-    7
-    >>> posify_index(10, [3, -3])
-    [3, 7]
-
-    >>> posify_index((10, 20), (3, -3))
-    (3, 17)
-    >>> posify_index((10, 20), (3, [3, 4, -3]))
-    (3, [3, 4, 17])
-    """
-    if isinstance(ind, tuple):
-        return tuple(map(posify_index, shape, ind))
-    if isinstance(ind, int):
-        if ind < 0:
-            return ind + shape
-        else:
-            return ind
-    if isinstance(ind, list):
-        return [posify_index(shape, i) for i in ind]
-    return ind
-
-
-slice_names = ('slice-%d' % i for i in count(1))
-
-
-def slice_array(out_name, in_name, blockdims, index):
-    """
-    Master function for array slicing
-
-    This function makes a new dask that slices blocks along every
-    dimension and aggregates (via cartesian product) each dimension's
-    slices so that the resulting block slices give the same results
-    as the original slice on the original structure
-
-    Parameters
-    ----------
-
-    in_name - string
-      This is the dask variable name that will be used as input
-    out_name - string
-      This is the dask variable output name
-    blockshape - iterable of integers
-    index - iterable of integres, slices, or lists
-
-    Returns
-    -------
-
-    Dict where the keys are tuples of
-
-        (out_name, dim_index[, dim_index[, ...]])
-
-    and the values are
-
-        (function, (in_name, dim_index, dim_index, ...),
-                   (slice(...), [slice()[,...]])
-
-    Also new blockdims
-
-    Example
-    -------
-
-    >>> slice_array('y', 'x', (100,), [(20, 20, 20, 20, 20)], (slice(10, 35),))  #  doctest: +SKIP
-    {('y', 0): (getitem, ('x', 0), (slice(10, 20),)),
-     ('y', 1): (getitem, ('x', 1), (slice( 0, 15),))}
-    """
-    index = tuple(index)
-    blockdims = tuple(map(tuple, blockdims))
-
-    if all(index == slice(None, None, None) for index in index):
-        return {out_name: in_name}, blockdims
-
-    missing = len(blockdims) - len([ind for ind in index if ind is not None])
-    index2 = index + (slice(None, None, None),) * missing
-    dsk_out, bd_out = slice_with_newaxes(out_name, in_name, blockdims, index2)
-
-    bd_out = tuple(map(tuple, bd_out))
-    return dsk_out, bd_out
-
-
-def insert_many(seq, where, val):
-    """ Insert value at many locations in sequence
-
-    >>> insert_many(['a', 'b', 'c'], [0, 2], 'z')
-    ('z', 'a', 'z', 'b', 'c')
-    """
-    seq = list(seq)
-    result = []
-    for i in range(len(where) + len(seq)):
-        if i in where:
-            result.append(val)
-        else:
-            result.append(seq.pop(0))
-    return tuple(result)
-
-
-def slice_with_newaxes(out_name, in_name, blockdims, index):
-    """
-    Handle indexing with Nones
-
-    Strips out Nones then hands off to slice_wrap_lists
-    """
-    assert all(isinstance(ind, (slice, int, list, type(None)))
-               for ind in index)
-
-    # Strip Nones from index
-    index2 = tuple([ind for ind in index if ind is not None])
-    where_none = [i for i, ind in enumerate(index) if ind is None]
-
-    # Pass down and do work
-    dsk, blockdims2 = slice_wrap_lists(out_name, in_name, blockdims, index2)
-
-    # Insert ",0" into the key:  ('x', 2, 3) -> ('x', 0, 2, 0, 3)
-    dsk2 = dict(((out_name,) + insert_many(k[1:], where_none, 0),
-                 (v[:2] + (insert_many(v[2], where_none, None),)))
-                for k, v in dsk.items()
-                if k[0] == out_name)
-
-    # Add back intermediate parts of the dask that weren't the output
-    dsk3 = merge(dsk2, dict((k, v) for k, v in dsk.items() if k[0] != out_name))
-
-    # Insert (1,) into blockdims:  ((2, 2), (3, 3)) -> ((2, 2), (1,), (3, 3))
-    blockdims3 = insert_many(blockdims2, where_none, (1,))
-
-    return dsk3, blockdims3
-
-
-def slice_wrap_lists(out_name, in_name, blockdims, index):
-    """
-    Fancy indexing along blocked array dasks
-
-    Handles index of type list.  Calls slice_slices_and_integers for the rest
-
-    See Also
-    --------
-
-    take - handle slicing with lists ("fancy" indexing)
-    slice_slices_and_integers - handle slicing with slices and integers
-    """
-    shape = tuple(map(sum, blockdims))
-    assert all(isinstance(i, (slice, list, int)) for i in index)
-
-    index2 = posify_index(shape, index)
-
-    # Do we have more than one list in the index?
-    where_list = [i for i, ind in enumerate(index) if isinstance(ind, list)]
-    if len(where_list) > 1:
-        raise NotImplementedError("Don't yet support nd fancy indexing")
-
-    # Turn x[5:10] into x[5:10, :, :] as needed
-    num_missing_dims = len(shape) - len([i for i in index2 if i is not None])
-    index3 = index2 + (slice(None, None, None),) * num_missing_dims
-
-    index_without_list = tuple(slice(None, None, None)
-                                    if isinstance(i, list)
-                                    else i
-                                    for i in index3)
-
-    # No lists, hooray! just use slice_slices_and_integers
-    if index3 == index_without_list:
-        return slice_slices_and_integers(out_name, in_name, blockdims, index3)
-
-    # lists and full slice/:   Just use take
-    if all(isinstance(i, list) or i == slice(None, None, None)
-            for i in index3):
-        axis = where_list[0]
-        dsk3 = take(out_name, in_name, blockdims, index3[where_list[0]],
-                    axis=axis)
-        blockdims2 = blockdims
-    else:  # Mixed case. Both slices/integers and lists. slice/integer then take
-        tmp = next(slice_names)
-        dsk, blockdims2 = slice_slices_and_integers(tmp, in_name, blockdims, index_without_list)
-
-        # After collapsing some axes due to int indices, adjust axis parameter
-        axis = where_list[0]
-        axis2 = axis - sum(1 for i, ind in enumerate(index3)
-                           if i < axis and isinstance(ind, int))
-
-        # Do work
-        dsk2 = take(out_name, tmp, blockdims2, index3[axis], axis=axis2)
-        dsk3 = merge(dsk, dsk2)
-
-    # Replace blockdims of list entries with single block
-    index4 = [ind for ind in index3 if not isinstance(ind, int)]
-    blockdims3 = tuple([bd if not isinstance(i, list) else (len(i),)
-                        for i, bd in zip(index4, blockdims2)])
-
-    return dsk3, blockdims3
-
-
-def slice_slices_and_integers(out_name, in_name, blockdims, index):
-    """
-    Dask array indexing with slices and integers
-
-    Assumes that index is only slice or int objects
-
-    See fancy_slice for full docstring
-
-    Returns
-    -------
-
-    New dask and new blockdims
-    """
-    shape = tuple(map(sum, blockdims))
-
-    assert all(isinstance(ind, (slice, int)) for ind in index)
-    assert len(index) == len(blockdims)
-
-    # Get a list (for each dimension) of dicts{blocknum: slice()}
-    block_slices = list(map(_slice_1d, shape, blockdims, index))
-
-    # (in_name, 1, 1, 2), (in_name, 1, 1, 4), (in_name, 2, 1, 2), ...
-    in_names = product([in_name], *[i.keys() for i in block_slices])
-
-    # (out_name, 0, 0, 0), (out_name, 0, 0, 1), (out_name, 0, 1, 0), ...
-    out_names = product([out_name],
-                        *[range(len(d))
-                            for d, i in zip(block_slices, index)
-                            if not isinstance(i, int)])
-
-    all_slices = list(product(*[i.values() for i in block_slices]))
-
-    dsk_out = dict((out_name, (getitem, in_name, slices))
-                   for out_name, in_name, slices
-                   in zip(out_names, in_names, all_slices))
-
-    new_blockdims = [new_blockdim(d, db, i)
-                     for d, i, db in zip(shape, index, blockdims)
-                     if not isinstance(i, int)]
-
-    return dsk_out, new_blockdims
-
-
 class Array(object):
     """ Array object holding a dask
